Add time counting, add it to db

+++ app/src/main/java/ee/taltech/sudoku/GameState.kt
@@ -4,6 +4,6 @@ class GameState(
     var id: Int,
     var gameBoard: String,
     var difficulty: String,
-    var timeSpent: Int,
+    var timeSpent: Long,
     var gameFinished: Int
 )
